Create More_function
Task 2: Write a function that moves a node forward in the given linked list. Take the pointer of the node to be moved as a parameter.
void moveforwardlist (struct node *, struct node *);

Task 3: Write a function that moves a node backward in the given linked list. Take the pointer of the node to be moved as a parameter.
void movebackwardlist (struct node *, struct node *);

Task 4: Write a function that loads the linked list with the given values of an array. Consider array and its size as input parameter to the function.
void loadlist (struct node *, int [ ], int );