Create More_Functions_1
Task 2: Write a function that finds the average of the elements of a  linked list. Use the following prototype. Modify your linked list data field to be a float and insert initial floats (1.5, 2.5, 3.5, 4.5) to test your function.
float averagelist (struct node *);


Task 3: Write a function that cosunts and returns 'A' grades which are stored in the linked list. Use the following prototype. Modify your linked list data field to be a character and insert initial characters (A,B,A,C,D) to test your function.
int countA (struct node *);